pyproject.toml
[tool.poetry]
name = "sessak-back"
version = "0.1.0"
description = ""
authors = ["dozzzini <dozzzini@naver.com>"]
readme = "README.md"
packages = [{include = "sessak_back"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
pillow = "^10.0.0"
django-allauth = "^0.56.1"
djangorestframework-simplejwt = "^5.3.0"
dj-rest-auth = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
